Move route components to separate folder
