+++ src/domain/app/routes/AppRoutes.tsx
@@ -2,12 +2,12 @@ import React from 'react';
 import { useTranslation } from 'react-i18next';
 import { Redirect, Route, Switch } from 'react-router';
 
-import { MobileMenuProvider } from '../../common/components/mobileMenu/MobileMenu';
-import { ResetFocus } from '../../common/route/RouteUtils';
-import { getCurrentLanguage } from '../../common/translation/TranslationUtils';
-import { SUPPORT_LANGUAGES } from '../../constants';
-import PageLayout from '../app/layout/PageLayout';
-import App from './App';
+import { MobileMenuProvider } from '../../../common/components/mobileMenu/MobileMenu';
+import { ResetFocus } from '../../../common/route/RouteUtils';
+import { getCurrentLanguage } from '../../../common/translation/TranslationUtils';
+import { SUPPORT_LANGUAGES } from '../../../constants';
+import PageLayout from '../layout/PageLayout';
+import App from './LocaleRoutes';
 
 const localeParam = `:locale(${SUPPORT_LANGUAGES.EN}|${SUPPORT_LANGUAGES.FI}|${SUPPORT_LANGUAGES.SV})`;
