+++ src/domain/app/routes/__tests__/LocaleRoutes.test.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import { withRouter } from 'react-router';
+
+import { render } from '../../../../util/testUtils';
+import LocaleRoutes from '../LocaleRoutes';
+
+const AppWithRouter = withRouter(LocaleRoutes);
+
+it('renders without crashing', () => {
+  render(<AppWithRouter />);
+});
