+++ src/domain/app/routes/constants.ts
@@ -1,0 +1,11 @@
+// Remember also update static urls to updateSitemap
+export enum ROUTES {
+  ABOUT = '/about',
+  ACCESSIBILITY = '/accessibility',
+  COLLECTION = '/collection/:slug',
+  COLLECTIONS = '/collections',
+  EVENTS = '/events',
+  EVENT = '/event/:id',
+  HOME = '/home',
+  HOME_PREVIEW = '/home/:id',
+}
