+++ src/index.tsx
@@ -2,8 +2,7 @@ import * as Sentry from '@sentry/browser';
 import React from 'react';
 import ReactDOM from 'react-dom';
 
-import BrowserApp from './domain/app/BrowserApp';
-import * as serviceWorker from './serviceWorker';
+import BrowserApp from './domain/app/App';
 
 if (process.env.REACT_APP_SENTRY_ENVIRONMENT) {
   Sentry.init({
@@ -13,8 +12,3 @@ if (process.env.REACT_APP_SENTRY_ENVIRONMENT) {
 }
 
 ReactDOM.hydrate(<BrowserApp />, document.getElementById('root'));
-
-// If you want your app to work offline and load faster, you can change
-// unregister() to register() below. Note this comes with some pitfalls.
-// Learn more about service workers: https://bit.ly/CRA-PWA
-serviceWorker.unregister();
